split osrm and valhalla route serializers
