+++ CMakeLists.txt
@@ -161,6 +161,10 @@ if(ENABLE_PYTHON_BINDINGS)
     set(Boost_Python_lib "Boost::python${PYTHON_VERSION_MAJOR}")
     set(_Boost_PYTHON${PYTHON_VERSION_MAJOR}_HEADERS "boost/python.hpp")
     find_package(Boost COMPONENTS python${PYTHON_VERSION_MAJOR})
+  elseif(APPLE AND ${PYTHON_VERSION_MAJOR} EQUAL 2 AND ${PYTHON_VERSION_MINOR} EQUAL 7)
+    set(Boost_Python_lib "Boost::python27")
+    set(_Boost_PYTHON27_HEADERS "boost/python.hpp")
+    find_package(Boost COMPONENTS python27)
   else()
     set(Boost_Python_lib "Boost::python")
     find_package(Boost COMPONENTS python)
